✨ refactor(buscador): implementar lógica de busca e resolução de captcha
Adiciona a classe BuscadorProcesso com métodos para buscar processos e resolver captchas no sistema PJe. Integra com tasks Celery e melhora o tratamento de exceções.

Files Changed:
- *crawjud/abstract/_buscador.py*: nova implementação da lógica de busca e resolução de captcha.
- *crawjud/abstract/_head.py*: nova classe HeadBot que herda de BuscadorProcesso.
- *crawjud/abstract/_master.py*: nova classe AbstractClassBot com métodos abstratos.
- *crawjud/bot.py*: ClassBot agora herda de HeadBot.
- *crawjud/bots/pje/capa.py*: ajusta chamada para buscar_processo.
- *crawjud/bots/pje/resources/buscador.py*: remoção do módulo obsoleto.
