+++ crawjud/abstract/_head.py
@@ -1,0 +1,4 @@
+from crawjud.abstract._buscador import BuscadorProcesso
+
+
+class HeadBot(BuscadorProcesso): ...
